+++ tabpro-ui/src/main/java/com/gstncaruso/tabpro/ui/score/ZoomHolder.java
@@ -1,0 +1,10 @@
+package com.gstncaruso.tabpro.ui.score;
+
+public interface ZoomHolder {
+
+    Zoom zoom();
+
+    void setZoom(Zoom zoom);
+
+    void onZoomChange(Runnable listener);
+}